requirements.txt
openai-whisper
flask
ffmpeg-python
